refactor: move proptypes to separated file
